+++ examples/mcpdft/h2co_tpbe66_def2svp_edipole.py
@@ -24,5 +24,5 @@ mc.kernel ()
 mc_dip = mc.electric_dipole()
 mol_dip = mc_dip.kernel ()
 print ("MC-PDFT electric dipole moment Debye \n {:8.5f} {:8.5f} {:8.5f}".format (*mol_dip))
-print ("Numerical MC-PDFT electric dipole moment from GAMESS [Debye] \n 2.0936 0.00000 0.00000 ")
+print ("Numerical MC-PDFT electric dipole moment from GAMESS [Debye] \n 2.09361 0.00000 0.00000 ")
